+++ packages/plugin-core/src/ExtensionProvider.ts
@@ -2,6 +2,7 @@ import { IDendronExtension } from "./dendronExtensionInterface";
 import { DendronError, DendronTreeViewKey } from "@dendronhq/common-all";
 import _ from "lodash";
 import { IWSUtilsV2 } from "./WSUtilsV2Interface";
+import { WorkspaceUtils } from "@dendronhq/engine-server";
 
 /**
  * Use this to statically get implementation of IDendronExtension without having to
@@ -38,6 +39,18 @@ export class ExtensionProvider {
 
   static isActive() {
     return ExtensionProvider.getExtension().isActive();
+  }
+
+  static isActiveAndIsDendronNote(fpath: string) {
+    if (!this.isActive()) {
+      return false;
+    }
+    const { wsRoot, vaults } = this.getDWorkspace();
+    return WorkspaceUtils.isDendronNote({
+      wsRoot,
+      vaults,
+      fpath,
+    });
   }
 
   static getWorkspaceConfig() {
